requirements.txt
ocrmypdf
pdf2image
pytesseract
Pillow
easyocr
opencv-python-headless
numpy
PyPDF2